requirements.txt
aiogram==3.24.0
discord.py==2.6.4
python-dotenv==1.2.1
